+++ changelog.md
@@ -1,5 +1,19 @@
 # Changelog for PSQuizmaster
 
+## v0.3.0
+
+### Changed
+
+- Changed parameter type for `Updated` in `Set-PSQuizFile` to `[datetime]` from `[string]`
+- Modified `Set-PSQuizFile` to accept pipeline input for the file name. Made the path mandatory
+
+### Added
+
+- Added parameter `Name` to `Get-PSQuizFile` to retrieve a quiz by name
+- Added alias `Start-PSQuiz` for `Invoke-PSQuiz`
+- Added command `New-PSQuiz` with an alias of `Make-PSQuiz`, to guide a quiz creator through the process of creating a quiz file with questions.
+- Added format file `psquizItem.format.ps1xml` to display quiz questions in a list format by default.
+
 ## v0.2.0
 
 ### Changed